requirements.txt
javalang==0.13.0
networkx==3.2.1
groq==0.13.0
rich==13.9.4
python-dotenv==1.0.1
